+++ src/helpers/vid.js
@@ -8,10 +8,10 @@ class VidHelper {
 
     /**
      * Save a DID document
-     * 
-     * @todo: Replace with decentralised lookup
+     *
+     * @todo: Replace with decentralised DID management (ie: sidetree)
      */
-    async save(did, appName, keyring, userDataserverUrl) {
+    async save(did, appName, keyring, userDataserverUrl, signature) {
         let vid = this.getVidFromDid(did, appName);
         let publicKeys = keyring.exportPublicKeys();
         let appUrl = App.config.appHost;
@@ -53,23 +53,31 @@ class VidHelper {
             serviceEndpoint: userDataserverUrl
         });
 
-        DIDHelper.createProof(doc, keyring.signKey.private);
-        let response = await DIDHelper.commit(did, doc, App.config.server.didServerUrl);
+        doc = DIDHelper.createProof(doc, keyring.signKey.private);
+
+        let response = await DIDHelper.commit(did, doc, signature, App.config.server.didServerUrl);
         if (response) {
             return doc;
         }
-
-        // Future: Have did-helper include consent message in the proof
-        // and have did-server verify the consent message is from a
-        // chain address that is linked to the DID (unless it's the only one)?
     }
 
+    /**
+     * Get a DID Document for the VID by DID and application name
+     * 
+     * @param {*} did Blockchain DID (ie: did:ethr:0x...)
+     * @param {*} appName Application name
+     */
     async getByDid(did, appName) {
         appName = appName || App.config.appName;
         did = did.toLowerCase();
         return await DIDHelper.loadForApp(did, appName, App.config.server.didServerUrl);
     }
 
+    /**
+     * Get a DID Document representing the VID
+     * 
+     * @param {string} vid VID to locate
+     */
     async getByVid(vid) {
         vid = vid.toLowerCase();
         return await DIDHelper.load(vid, App.config.server.didServerUrl);
@@ -77,7 +85,7 @@ class VidHelper {
 
     /**
      * Get DID for a given VID
-     * 
+     *
      * @param {*} vid
      */
     async getDidFromVid(vid) {
@@ -87,19 +95,66 @@ class VidHelper {
 
     /**
      * Get the VID for a given DID and application name
-     * 
-     * @param {*} did 
-     * @param {*} appName 
+     *
+     * @param {*} did
+     * @param {*} appName
      */
     getVidFromDid(did, appName) {
         appName = appName || App.config.appName;
         did = did.toLowerCase();
-        return 'did:vid:' + utils.id(appName + did);
+        return 'did:verida:' + utils.id(appName + did);
     }
 
+    /**
+     * Convert a VID to the underlying DID and application name
+     * 
+     * @param {*} vid 
+     */
+    async convertVid(vid) {
+        const didDoc = await this.getByVid(vid);
+        const applicationService = didDoc.service.find(entry => entry.type.includes("verida.Application"));
+        const appName = applicationService.description;
+        const did = await this.getDidFromVid(vid);
+        
+        return {
+            did: did,
+            appName: appName
+        };
+    }
+
+    /**
+     * Construct a DID given a chain and address
+     * 
+     * @param {*} address 
+     * @param {*} chain 
+     */
     getDidFromAddress(address, chain) {
         chain = chain || "ethr";
         return 'did:'+chain+':'+address.toLowerCase();
+    }
+
+    /**
+     * Convert a username to a DID
+     * 
+     */
+    async getDidFromUsername(username) {
+        if (usernameOrDid.match(/^did\:/)) {
+            return username;
+        }
+
+        return DIDHelper.getDidFromUsername(username, App.config.server.didServerUrl);
+    }
+
+    /**
+     * Save a username for a given DID. Requires a valid signature signed
+     * by the DID.
+     * 
+     * @param {*} username 
+     * @param {*} did 
+     * @param {*} signature 
+     */
+    async commitUsername(username, did, signature) {
+        return DIDHelper.commitUsername(username, did, signature, App.config.server.didServerUrl);
     }
 
 }
@@ -113,6 +168,6 @@ export function getResolver() {
     }
 
     return {
-        vid: resolve
+        verida: resolve
     }
 }