+++ src/keyring.js
@@ -1,17 +1,13 @@
 /*eslint no-console: "off"*/
-import { secretbox, box, sign, randomBytes } from "tweetnacl";
+import { box, sign } from "tweetnacl";
 import {
-  decodeUTF8,
-  encodeUTF8,
-  encodeBase64,
-  decodeBase64
+  encodeBase64
 } from "tweetnacl-util";
-import { utils, ethers } from 'ethers';
+import { ethers } from 'ethers';
+import Encryption from "@verida/encryption-utils";
 
-const newSymNonce = () => randomBytes(secretbox.nonceLength);
-const newAsymNonce = () => randomBytes(box.nonceLength);
-
-const BASE_PATH = "m/6696500'/0'/0'"
+const BASE_PATH = "m/6696500'/0'/0'";
+const DB_PATH = "m/42"
 //const ETH_PATH = "m/44'/60'/0'/0"
 
 /**
@@ -28,26 +24,48 @@ class Keyring {
     constructor(signature) {
         this.signature = signature;
 
-        const entropy = utils.sha256('0x' + signature.slice(2));
-        const mnemonic = ethers.HDNode.entropyToMnemonic(entropy);
-        const seed = ethers.HDNode.mnemonicToSeed(mnemonic);
-
-        const seedNode = ethers.HDNode.fromSeed(seed);
-        const baseNode = seedNode.derivePath(BASE_PATH);
+        const entropy = ethers.utils.sha256('0x' + signature.slice(2));
+        const seedNode = ethers.utils.HDNode.fromSeed(entropy);
+        this.baseNode = seedNode.derivePath(BASE_PATH);
 
         // Build symmetric key
-        let symKey = baseNode.derivePath("0");
+        let symKey = this.baseNode.derivePath("0");
         this.symKey = Buffer.from(symKey.privateKey.slice(2), 'hex');
 
         // Build asymmetric keys
-        let asymKey = baseNode.derivePath("1");
+        let asymKey = this.baseNode.derivePath("1");
         this.asymKey = this._generateKeyPair(asymKey, "box");
 
         // Build signing keys
-        let signKey = baseNode.derivePath("2");
+        let signKey = this.baseNode.derivePath("2");
         this.signKey = this._generateKeyPair(signKey, "sign");
 
+        const dbNode = this.baseNode.derivePath("3");
+        this.dbSignKey = this._generateKeyPair(dbNode, "sign");
+        this.dbSymKeys = {};
+
         //this.ethKey = seedNode.derivePath(ETH_PATH);
+    }
+
+    getDbKey(dbName) {
+        if (this.dbSymKeys[dbName]) {
+            return this.dbSymKeys[dbName];
+        }
+
+        // Sign a consent message using the current db signing key
+        const consent = "Authorized to own database: " + dbName;
+        const signature = this.sign(consent, this.dbSignKey);
+        const signatureBytes = ethers.utils.toUtf8Bytes(signature)
+
+        // Use the signature as entropy to create a new seed
+        const entropy = ethers.utils.keccak256(signatureBytes);
+        const seedNode = ethers.utils.HDNode.fromSeed(entropy);
+        const dbNode = seedNode.derivePath(DB_PATH);
+
+        // Use the HDNode to create a symmetric key for this database
+        this.dbSymKeys[dbName] = dbNode.privateKey;
+
+        return this.dbSymKeys[dbName];
     }
 
     exportPublicKeys() {
@@ -85,111 +103,37 @@ class Keyring {
         };
     }
 
-    // get a signature
+    // Sign arbitrary data
     sign(data) {
-        let messageUint8 = decodeUTF8(JSON.stringify(data));
-        return encodeBase64(sign.detached(messageUint8, this.signKey.privateBytes));
+        return Encryption.signData(data, this.signKey.privateBytes)
     }
 
     verifySig(data, sig) {
-        let messageUint8 = decodeUTF8(JSON.stringify(data));
-        return sign.detached.verify(messageUint8, decodeBase64(sig), this.signKey.publicBytes);
+        return Encryption.verifySig(data, sig, this.signKey.publicBytes)
     }
 
     symEncryptBuffer(data) {
-        const keyUint8Array = this.symKey;
-
-        const nonce = newSymNonce();
-        const messageUint8 = data;
-        const box = secretbox(messageUint8, nonce, keyUint8Array);
-
-        const fullMessage = new Uint8Array(nonce.length + box.length);
-        fullMessage.set(nonce);
-        fullMessage.set(box, nonce.length);
-
-        const base64FullMessage = encodeBase64(fullMessage);
-        return base64FullMessage;
+        return Encryption.symEncryptBuffer(data, this.symKey);
     }
 
     symDecryptBuffer(messageWithNonce) {
-        const keyUint8Array = this.symKey;
-        const messageWithNonceAsUint8Array = decodeBase64(messageWithNonce);
-        const nonce = messageWithNonceAsUint8Array.slice(0, secretbox.nonceLength);
-        const message = messageWithNonceAsUint8Array.slice(
-            secretbox.nonceLength,
-            messageWithNonce.length
-        );
-
-        const decrypted = secretbox.open(message, nonce, keyUint8Array);
-        if (!decrypted) {
-            throw new Error("Could not decrypt message");
-        }
-
-        return decrypted;
+        return Encryption.symDecryptBuffer(messageWithNonce, this.symKey);
     }
 
     symEncrypt(data) {
-        const keyUint8Array = this.symKey;
-
-        const nonce = newSymNonce();
-        const messageUint8 = decodeUTF8(JSON.stringify(data));
-        const box = secretbox(messageUint8, nonce, keyUint8Array);
-
-        const fullMessage = new Uint8Array(nonce.length + box.length);
-        fullMessage.set(nonce);
-        fullMessage.set(box, nonce.length);
-
-        const base64FullMessage = encodeBase64(fullMessage);
-        return base64FullMessage;
+        return Encryption.symEncrypt(data, this.symKey);
     }
 
     symDecrypt(messageWithNonce) {
-        const keyUint8Array = this.symKey;
-        const messageWithNonceAsUint8Array = decodeBase64(messageWithNonce);
-        const nonce = messageWithNonceAsUint8Array.slice(0, secretbox.nonceLength);
-        const message = messageWithNonceAsUint8Array.slice(
-            secretbox.nonceLength,
-            messageWithNonce.length
-        );
-
-        const decrypted = secretbox.open(message, nonce, keyUint8Array);
-        if (!decrypted) {
-            throw new Error("Could not decrypt message");
-        }
-
-        const base64DecryptedMessage = encodeUTF8(decrypted);
-        return JSON.parse(base64DecryptedMessage);
+        return Encryption.symDecrypt(messageWithNonce, this.symKey);
     }
 
     asymEncrypt(data, secretOrSharedKey) {
-        const nonce = newAsymNonce();
-        const messageUint8 = decodeUTF8(JSON.stringify(data));
-        const encrypted = box.after(messageUint8, nonce, secretOrSharedKey);
-
-        const fullMessage = new Uint8Array(nonce.length + encrypted.length);
-        fullMessage.set(nonce);
-        fullMessage.set(encrypted, nonce.length);
-
-        const base64FullMessage = encodeBase64(fullMessage);
-        return base64FullMessage;
+        return Encryption.asymEncrypt(data, secretOrSharedKey);
     }
 
     asymDecrypt(messageWithNonce, secretOrSharedKey) {
-        const messageWithNonceAsUint8Array = decodeBase64(messageWithNonce);
-        const nonce = messageWithNonceAsUint8Array.slice(0, box.nonceLength);
-        const message = messageWithNonceAsUint8Array.slice(
-            box.nonceLength,
-            messageWithNonce.length
-        );
-
-        const decrypted = box.open.after(message, nonce, secretOrSharedKey);
-
-        if (!decrypted) {
-            throw new Error('Could not decrypt message');
-        }
-
-        const base64DecryptedMessage = encodeUTF8(decrypted);
-        return JSON.parse(base64DecryptedMessage);
+        return Encryption.asymDecrypt(messageWithNonce, secretOrSharedKey);
     }
 
     buildSharedKeyStart(privateKey) {