+++ src/user/server.js
@@ -1,4 +1,5 @@
-import { ecsign, hashPersonalMessage, toRpcSig } from 'ethereumjs-util';
+import utils from '@verida/wallet-utils';
+
 import Base from './base';
 
 class ServerUser extends Base {
@@ -10,6 +11,8 @@ class ServerUser extends Base {
      *
      * @property {string} did Decentralised ID for this use (ie: `did:ethr:0xaef....`)
      * @property {string} address Blockchain address for this user (ie: `0xaef....`)
+     * @property {string} appServerUrl URL of the Verida app server 
+     * @property {string} privateKey Private key (with leading `0x`)
      */
     constructor(chain, address, appServerUrl, privateKey) {
         super(chain, address, appServerUrl);
@@ -18,16 +21,12 @@ class ServerUser extends Base {
             throw new Error("No private key specified for server user");
         }
 
-        const key = privateKey.slice(2);
-        this.privateKey = Buffer.from(key, 'hex');
+        this.privateKeyHex = privateKey;
+        this.privateKey = Buffer.from(privateKey.slice(2), 'hex');
     }
 
-    async requestSignature(appName, accessType) {
-        let signMessage = this._getSignMessage(appName, accessType);
-        let message = Buffer.from(signMessage);
-        let messageHash = hashPersonalMessage(message);
-        let sig = ecsign(messageHash, this.privateKey);
-        return toRpcSig(sig.v, sig.r, sig.s);
+    async _requestSignature(signMessage) {
+        return utils.signMessage(this.chain, this.privateKeyHex, signMessage)
     }
 
 }